GO-Fish
A terminal version of go fish. The names of the players are needed, the book size is needed. 

The deck gets shuffled. The cards delt. and play begins, scoring is automatic. 

There is no lying in this version of go fish. But it is very easy to see other peoples cards since the only print out is to the terminal in version 1.